# [crypto/ec] constant time P-384 EC_METHOD