--- a/crypto/ec/ec_local.h
+++ b/crypto/ec/ec_local.h
@@ -614,6 +614,7 @@ const EC_METHOD *EC_GFp_s390x_nistp256_method(void);
 const EC_METHOD *EC_GFp_s390x_nistp384_method(void);
 const EC_METHOD *EC_GFp_s390x_nistp521_method(void);
 #endif
+const EC_METHOD *ec_GFp_secp384r1_method(void);
 
 size_t ec_key_simple_priv2oct(const EC_KEY *eckey,
                               unsigned char *buf, size_t len);