We've been experimenting with tooling (edit: now public, called [ECCKiila](https://gitlab.com/nisec/ecckiila) with [preprint](https://arxiv.org/abs/2007.11481)) that generates curve-specific `EC_METHOD`s. The idea is you get a whole stack similar to `ecp_nistp224.c` `ecp_nistp256.c` `ecp_nistp521.c`. Features include

* 100% computer generated, including the OpenSSL rigging
* Most of the GF layer comes from [fiat](https://github.com/mit-plv/fiat-crypto), so you get some formal verification guarantees from there (which is where the guarantees end)
* CT scalar multiplication by base point (keygen, sign)
* CT scalar multiplication by arbitrary point (derive) 
 static
* variable time scalar multiplication for two points (verify)
* Both 64-bit (highly tested) and 32-bit support (not so much)
* No alignment assumptions (not explicitly tested)
* No endianness assumptions (coded that way, yet untested on BE)

~~So this is just a WiP PR to solicit feedback. Also, merge #12096 then I'll rebase and we should see CI failures ;)~~ out of WiP with e84eaffc4908f6a926ec8fec264fa14ee3bc1ccd

It is really push-button stuff so whatever curve you think is important for the lib, it takes like 2min.

P-384 isn't the best curve to showcase the value of the tool because of size and limited GF optimization options. But nevertheless, maybe it's important for the lib because it's currently a gap in FIPS coverage.

Tagging @mattcaswell @romen 

## Before

```
$ apps/openssl speed ecdsap384
Doing 384 bits sign ecdsa's for 10s: 9611 384 bits ECDSA signs in 10.00s 
Doing 384 bits verify ecdsa's for 10s: 11739 384 bits ECDSA verify in 10.00s
version: 3.0.0-alpha4-dev
built on: built on: Fri Jun 19 06:56:42 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache clang-10 -fPIC -pthread -m64 -Wa,--noexecstack -Qunused-arguments -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              sign    verify    sign/s verify/s
 384 bits ecdsa (nistp384)   0.0010s   0.0009s    961.1   1173.9
$ apps/openssl speed ecdhp384
Doing 384 bits  ecdh's for 10s: 10057 384-bits ECDH ops in 10.00s
version: 3.0.0-alpha4-dev
built on: built on: Fri Jun 19 06:56:42 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache clang-10 -fPIC -pthread -m64 -Wa,--noexecstack -Qunused-arguments -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              op      op/s
 384 bits ecdh (nistp384)   0.0010s   1005.7
```

## After
```
$ apps/openssl speed ecdsap384
Doing 384 bits sign ecdsa's for 10s: 41855 384 bits ECDSA signs in 10.00s 
Doing 384 bits verify ecdsa's for 10s: 11740 384 bits ECDSA verify in 10.00s
version: 3.0.0-alpha4-dev
built on: built on: Tue Jun 30 05:10:45 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache clang -fPIC -pthread -m64 -Wa,--noexecstack -Qunused-arguments -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              sign    verify    sign/s verify/s
 384 bits ecdsa (nistp384)   0.0002s   0.0009s   4185.5   1174.0

$ apps/openssl speed ecdhp384
Doing 384 bits  ecdh's for 10s: 13119 384-bits ECDH ops in 10.00s
version: 3.0.0-alpha4-dev
built on: built on: Tue Jun 30 05:10:45 2020 UTC
options:bn(64,64) rc4(16x,int) des(int) aes(partial) idea(int) blowfish(ptr) 
compiler: ccache clang -fPIC -pthread -m64 -Wa,--noexecstack -Qunused-arguments -Wall -O3 -DOPENSSL_USE_NODELETE -DL_ENDIAN -DOPENSSL_PIC -DOPENSSL_BUILDING_OPENSSL -DNDEBUG
CPUINFO: OPENSSL_ia32cap=0x7ffef3ffffebffff:0x18d39efffb
                              op      op/s
 384 bits ecdh (nistp384)   0.0008s   1311.9
```

~~The slight regression on ECDSA verify is because we haven't really put any effort there, and also it speaks to the efficiency of the old NAF code in OpenSSL.~~ more performance improvements in 7e10f3b7cbb0b5f0d2ae553af4a281da80478ccf