@bbbrumley 2020-06-30
After FOO goes in, I will have to do silliness like this:

https://github.com/openssl/openssl/blob/master/crypto/ec/ecp_nistp521.c#L1914-L1926

A better approach would be, inside `EC_GROUP`, consolidating `a_is_minus3` and `asn1_flag` into a general `flags` field, then reserve a bit for dirty generators.