@romen 2020-06-30
> After #12096 goes in, I will have to do silliness like this:
> 
> https://github.com/openssl/openssl/blob/master/crypto/ec/ecp_nistp521.c#L1914-L1926
> 
> A better approach would be, inside `EC_GROUP`, consolidating `a_is_minus3` and `asn1_flag` into a general `flags` field, then reserve a bit for dirty generators.

Yeah, actually the `asn1_flag` field is a bit meh right now in `master` anyway, as it is used by legacy code but not by providers, and I would say it is not correct to have it inside `EC_GROUP` as it should be a `EVP_PKEY`/`EC_KEY` parameter or a `OSSL_SERIALIZER_CTX` option to select if a particular key should be encoded with explicit parameters or using the curve OID upon serialization.

This is a discussion that deserves its own issue, but my recommendation if you decide to consolidate the flags is to leave `asn1_flag` aside as an exception, as it is controversial and will likely slow down the development of unrelated features.