@romen 2020-07-01
> I don't really understand the latest Travis failure. How can it pass on `CONFIG_OPTS="--strict-warnings"` but fail on `CONFIG_OPTS="--strict-warnings" MAKEVERBOSE="yes"` ?

The failure is happening on ARM64, I can't look into details now and I have very limited availability until Sunday.