@bbbrumley 2020-07-12
> The failure is happening on ARM64

This is now [fixed](https://gitlab.com/nisec/ecckiila/-/commit/809a3310ad142dde8e89afa81d3ab6fa9de3a164). `char` defaults to `signed` on x86/x64, but `signed` on ARM.