@bbbrumley 2020-07-12
Rebased in 80dccc62c0fb6da37326f1b35b5fc41ca79a4127 (thanks @beldmit for merging #12096). We can see the test failure now:

```
    # Subtest: custom_generator_test
    1..82
    # Curve secp112r1
    ok 1 - iteration 1
...
    # Curve secp224r1
    ok 10 - iteration 10
    # Curve secp256k1
    ok 11 - iteration 11
    # Curve secp384r1
    # ERROR: (int) 'CRYPTO_memcmp(b1, b2, bsize) == 0' failed @ test/ectest.c:2405
    # [1] compared to [0]
    not ok 12 - iteration 12
```

I will start working on the fix, then I think I can take this out of WiP.