@bbbrumley 2020-07-12
> Do you mean unsigned on s390x?

I meant unsigned on armv7 and aarch64 :) This code should rarely get compiled in on s390x.