@kroeckx 2020-07-12
I forgot that arm was unsigned too. powerpc would be the other that's unsigned.