@kroeckx 2020-07-25
It was not clear that you took this out of WIP. Are you waiting for a review now?

I see that NSS has just released a version with P384 and P521 based on this.