@bbbrumley 2020-07-26
> Are you waiting for a review now?

From my perspective, yea it's ready :)