@bbbrumley 2020-07-26
Also [the preprint](https://arxiv.org/abs/2007.11481) is now live. Table 2 gives some benchmarks.