@beldmit 2020-07-28
It will be very pity if this great work is not included in the OpenSSL because of the licence clash.