@mattcaswell 2020-08-26
>  my current understanding is that MIT doesn't permit us to sign CLA's, so I don't think you'll be able to get CLAs for fiat-crypto

We had an OMC meeting today and discussed this quite a bit. Ultimately the OMC decided that we cannot make an exception to our CLA rules and so, unfortunately, this PR is not going to be acceptable.

This is a great shame as I can see the value that having this code would bring (and I think that sentiment was shared by the rest of the OMC). I'd encourage looking at making this code available via a third party provider.