@bbbrumley 2020-08-28
> this PR is not going to be acceptable.

That's unfortunate.

If you change you mind, ECCKiila is there and you can generate whatever stack you want for any curve, not just P-384.

So if there's a next time, you won't need me or my PR ;)