@DDvO 2020-09-29
I'm just copying @mattcaswell's most recent reply in the discussion on this PR 
from a within a sub-thread shown further up https://github.com/openssl/openssl/pull/12201#discussion_r491881293 down to here
such that the current state of the discussion becomes easier to find/see:

> Well there are two issues: licence and CLAs. 
Your offer above to relicence to Apache 2.0 solves the licence issue. 
The CLA issue is independent of licensing. 
So, unfortunately, I don't think swapping to a different licence solves it.